Replace load_task_from_disk with the new format and move to serialization module
